requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyzmq==25.1.1
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.2 